# [Quant] Improve performance of ONEDNN backend