--- a/aten/src/ATen/native/quantized/cpu/qconv.cpp
+++ b/aten/src/ATen/native/quantized/cpu/qconv.cpp
@@ -1227,49 +1227,89 @@ at::Tensor PackedConvWeightsOnednn<kSpatialDim>::apply_impl(
   const ideep::dims& dilates = dilation().vec();
   const ideep::dims& padding_l = padding().vec();
   const ideep::dims& padding_r = padding().vec();
-  const ideep::scale_t& src_scales = ideep::scale_t(1, 1.0/act.q_scale()); // Scales of ONEDNN and PyTorch are reciprocal
+  double input_scale = act.q_scale();
+  int64_t input_zp = act.q_zero_point();
+  // Scales of ONEDNN and PyTorch are reciprocal
+  const ideep::scale_t& src_scales = ideep::scale_t(1, 1.0/input_scale);
   const ideep::scale_t& weights_scales = weights.get_scale();
-  const ideep::scale_t& dst_scales = ideep::scale_t(weights_scales.size(), 1.0/output_scale); // Scales of ONEDNN and PyTorch are reciprocal
-  const ideep::zero_point_t src_zero_points = ideep::zero_point_t(1, act.q_zero_point());
+  int64_t scale_size = weights_scales.size();
+  double inv_output_scale = 1.0/output_scale;
+  const ideep::zero_point_t src_zero_points = ideep::zero_point_t(1, input_zp);
   const ideep::zero_point_t dst_zero_points = ideep::zero_point_t(1, output_zero_point);
   ideep::attr_t op_attr = kReluFused ? ideep::attr_t::fuse_relu() : ideep::attr_t();
-  op_attr.set_zero_points(DNNL_ARG_SRC, ideep::utils::tensor_zp_mask(1), {DNNL_RUNTIME_S32_VAL}); // runtime src zero point
-  if (with_bias) {
-    // Bias might be modified outside (e.g. by quantization bias correction).
-    // If so, update the prepacked bias as well.
-    if (bias_.value().get_data_handle() != orig_bias_.value().data_ptr()) {
-      bias_.value().init(bias_.value().get_desc(), orig_bias_.value().data_ptr());
-    }
-    const auto& b = bias_.value();
-    if (transpose()) {
-      ideep::convolution_transpose_forward::compute_v2(
+  // Since src zero point is unknown, set runtime value here
+  op_attr.set_zero_points(DNNL_ARG_SRC, ideep::utils::tensor_zp_mask(1), {DNNL_RUNTIME_S32_VAL});
+
+  // Bias might be modified outside (e.g. by quantization bias correction).
+  // If so, update the prepacked bias as well.
+  if (with_bias && bias_.value().get_data_handle() != orig_bias_.value().data_ptr()) {
+    bias_.value().init(bias_.value().get_desc(), orig_bias_.value().data_ptr());
+  }
+  const auto& b = with_bias ? bias_.value() : ideep::tensor();
+  int num_threads = at::get_num_threads();
+  if (transpose()) {
+    // Primitive cache is initialized when called for the first time
+    // and won't be updated afterwards.
+    PrimitiveCacheKey cache_key = std::make_tuple(
+        input_scale, input_zp, src_dims, output_scale, output_zero_point, num_threads);
+    std::call_once(*cache_initialized_flag, [&](){
+        DeconvParams params;
+        ideep::convolution_transpose_forward::prepare(
+            params, src, weights, b, dst_dims, dst,
+            strides, padding_l, padding_r, dilates, groups(),
+            src_scales, weights_scales, ideep::scale_t(scale_size, inv_output_scale),
+            src_zero_points, dst_zero_points, op_attr,
+            dnnl::algorithm::deconvolution_direct,
+            dnnl::prop_kind::forward_inference,
+            ideep::u8s8, ideep::engine::cpu_engine());
+        get_deconv_cache() = DeconvPrimitiveCache(cache_key, params, b);
+        weights = weights.reorder_if_differ_in(params.pd.weights_desc());
+    });
+    if (get_deconv_cache().hit(cache_key)) {
+      DeconvParams& params = get_deconv_cache().get_params();
+      auto& expected_bias = get_deconv_cache().get_bias();
+      ideep::convolution_transpose_forward::compute<false, false>(
+          params, src, weights, expected_bias, dst);
+    } else {
+      ideep::convolution_transpose_forward::compute(
           src, weights, b, dst_dims, dst,
           strides, padding_l, padding_r, dilates,
-          groups(), src_scales, weights_scales, dst_scales, src_zero_points, dst_zero_points,
-          op_attr, dnnl::algorithm::deconvolution_direct, dnnl::prop_kind::forward_inference,
-          ideep::u8s8, ideep::engine::cpu_engine());
-    } else {
-      ideep::convolution_forward::compute_v2(
-          src, weights, b, dst_dims, dst,
-          strides, dilates, padding_l, padding_r, groups(),
-          src_scales, weights_scales, dst_scales, src_zero_points, dst_zero_points,
-          op_attr, dnnl::algorithm::convolution_direct, dnnl::prop_kind::forward_inference,
+          groups(), src_scales, weights_scales,
+          ideep::scale_t(scale_size, inv_output_scale),
+          src_zero_points, dst_zero_points, op_attr,
+          dnnl::algorithm::deconvolution_direct,
+          dnnl::prop_kind::forward_inference,
           ideep::u8s8, ideep::engine::cpu_engine());
     }
-  } else {
-    if (transpose()) {
-      ideep::convolution_transpose_forward::compute_v2(
-          src, weights, dst_dims, dst,
-          strides, padding_l, padding_r, dilates,
-          groups(), src_scales, weights_scales, dst_scales, src_zero_points, dst_zero_points,
-          op_attr, dnnl::algorithm::deconvolution_direct, dnnl::prop_kind::forward_inference,
-          ideep::u8s8, ideep::engine::cpu_engine());
+  } else {  // not transposed
+    PrimitiveCacheKey cache_key = std::make_tuple(
+        input_scale, input_zp, src_dims, output_scale, output_zero_point, num_threads);
+    std::call_once(*cache_initialized_flag, [&](){
+        ConvParams params;
+        ideep::convolution_forward::prepare(
+            params, src, weights, b, dst_dims, dst,
+            strides, dilates, padding_l, padding_r, groups(),
+            src_scales, weights_scales, ideep::scale_t(scale_size, inv_output_scale),
+            src_zero_points, dst_zero_points,
+            op_attr, dnnl::algorithm::convolution_direct,
+            dnnl::prop_kind::forward_inference,
+            ideep::u8s8, ideep::engine::cpu_engine());
+        get_conv_cache() = ConvPrimitiveCache(cache_key, params, b);
+        weights = weights.reorder_if_differ_in(params.pd.weights_desc());
+    });
+    // If hit, use cached data. If miss, fall back to normal path.
+    if (get_conv_cache().hit(cache_key)) {
+      auto& params = get_conv_cache().get_params();
+      auto& expected_bias = get_conv_cache().get_bias();
+      ideep::convolution_forward::compute<false, false>(params, src, weights, expected_bias, dst);
     } else {
-      ideep::convolution_forward::compute_v2(
-          src, weights, dst_dims, dst,
+      ideep::convolution_forward::compute(
+          src, weights, b, dst_dims, dst,
           strides, dilates, padding_l, padding_r, groups(),
-          src_scales, weights_scales, dst_scales, src_zero_points, dst_zero_points,
-          op_attr, dnnl::algorithm::convolution_direct, dnnl::prop_kind::forward_inference,
+          src_scales, weights_scales, ideep::scale_t(scale_size, inv_output_scale),
+          src_zero_points, dst_zero_points, op_attr,
+          dnnl::algorithm::convolution_direct,
+          dnnl::prop_kind::forward_inference,
           ideep::u8s8, ideep::engine::cpu_engine());
     }
   }